+++ src/view-models/IPageSwitchViewModel.ts
@@ -1,0 +1,9 @@
+import React, { Dispatch, SetStateAction } from 'react'
+
+
+export interface IPageSwitchViewModel {
+	changePage(e: React.MouseEvent,
+			   setCurrentPage: Dispatch<SetStateAction<number>>
+	): void
+
+}
